agents/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
httpx